move filesystem-based implementations to 'fs' package
